crates: lofire-repo -> lofire
