+++ lofire-broker/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 
 [dependencies]
 lofire = { path = "../lofire" }
-lofire-repo = { path = "../lofire-repo" }
 blake3 = "1.3.1"
 chacha20 = "0.9.0"
 serde = { version = "1.0", features = ["derive"] }
