+++ lofire-broker/src/types.rs
@@ -3,7 +3,6 @@
 //! Corresponds to the BARE schema
 
 use lofire::types::*;
-use lofire_repo::types::*;
 use serde::{Deserialize, Serialize};
 
 //
