+++ lofire-p2p/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 lofire = { path = "../lofire" }
 lofire-broker = { path = "../lofire-broker" }
-lofire-repo = { path = "../lofire-repo" }
 blake3 = "1.3.1"
 chacha20 = "0.9.0"
 serde = { version = "1.0", features = ["derive"] }
