+++ lofire-p2p/src/types.rs
@@ -4,7 +4,6 @@
 
 use lofire::types::*;
 use lofire_broker::types::*;
-use lofire_repo::types::*;
 use serde::{Deserialize, Serialize};
 
 //
