+++ lofire/src/lib.rs
@@ -1,1 +1,13 @@
 pub mod types;
+
+pub mod store;
+
+pub mod tree;
+
+pub mod object;
+
+pub mod commit;
+
+pub mod branch;
+
+pub mod repo;
